backend/requirements.txt
aiosmtpd==1.4.6
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
arabic-reshaper==3.0.0
atpublic==6.0.1
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
httpx==0.27.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pillow==11.3.0
pillow-heif==0.18.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-bidi==0.6.6
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.4.4
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
websockets==15.0.1
openpyxl==3.1.5
segno==1.6.6
sse-starlette==1.8.2
weasyprint==62.3
pydyf==0.11.0
chromadb==0.5.23
rq==1.16.2
redis==5.0.8
